up wrapped components
