+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/Generic.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7b88990938634536b2c1172b4ba33af3
+timeCreated: 1687355367
