+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/Generic/IWrappedComponent.cs
@@ -1,0 +1,7 @@
+namespace DCL.ECS7.ComponentWrapper.Generic
+{
+    public interface IWrappedComponent<T> : IWrappedComponent where T: class
+    {
+        T Model { get; }
+    }
+}
