+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/Generic/IWrappedComponent.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c3296d588b6947d9ac1cc17958a6ba49
+timeCreated: 1687355375
