+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/Generic/PooledWrappedComponent.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f701441ee6494606ab29f25ea7f0a44e
+timeCreated: 1687355543
