+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/IPooledWrappedComponent.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace DCL.ECS7.ComponentWrapper
+{
+    public interface IPooledWrappedComponent : IDisposable
+    {
+        IWrappedComponent WrappedComponentBase { get; }
+    }
+}
