+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/IPooledWrappedComponent.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ff960cb3a05b42f5a313afb52382d7a3
+timeCreated: 1687355463
