+++ unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/IWrappedComponent.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 6d356afaffee414bb2aaf807ece16314
+timeCreated: 1687355332
