--- unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/WrappedComponent.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: dcda99e3c90e46439855ecaa5b4f4c4e
-timeCreated: 1687281529
