--- unity-renderer/Assets/DCLPlugins/ECS7/ComponentWrapper/WrappedComponentPool.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 2edae68b5fae4f1f8967f0eb95cd75d2
-timeCreated: 1687283062
